requirements.txt

streamlit
google-generativeai
google-adk
